Refactor change info handling in Chrononaut. Add RecordChanges mixin.
