+++ tests/conftest.py
@@ -148,7 +148,11 @@ def generate_test_models(db):
         roles = db.relationship('Role', secondary=roles_users,
                                 backref=db.backref('users', lazy='dynamic'))
 
-    return Todo, UnversionedTodo, SpecialTodo, Report, User, Role
+    class ChangeLog(db.Model, chrononaut.change_info.RecordChanges):
+        id = db.Column(db.Integer, primary_key=True)
+        note = db.Text()
+
+    return Todo, UnversionedTodo, SpecialTodo, Report, User, Role, ChangeLog
 
 
 @pytest.yield_fixture(scope='function')
